fix(astro): avoid pulling the SSR runtime graph into the content layer
The reference-validation code made content-layer.ts import isContentReference
from runtime.js, which transitively pulls the whole SSR runtime module graph
(assets/runtime.js, runtime/server/index.js, ...) into the build/sync module
graph for the first time. That extra graph in the dev server's module context
was breaking content-collection image HMR recovery on CI
(content-collections-image-hmr.test.ts: 500 instead of 200), even though the
affected fixture uses image() and never calls reference().

Move REFERENCE_MARKER / markReference / isContentReference into a small
dependency-free module (reference-marker.ts) that both runtime.ts and
content-layer.ts import, so the content layer no longer depends on runtime.js.
No behavior change to reference validation.
